Remove disconnect InboundCommand
Now the way to signal disconnect is to send and EOF on the next read request. This is nice as it removes a second way of signaling the channel is closed mostly because the tail of the stage is really the driver of the pipeline so it should always be attempting to read if its ready for commands anyway.